+++ modules/space/rendering/renderablesmallbody.cpp
@@ -202,7 +202,10 @@ void RenderableSmallBody::readDataFile(const std::string& filename) {
 
         unsigned int sequentialLineErrors = 0;
         unsigned int endElement = _startRenderIdx + _sizeRender - 1;
-        endElement = (endElement >= _numObjects) ? _numObjects - 1 : endElement;
+        endElement =
+            (endElement >= _numObjects) ?
+            static_cast<unsigned int>(_numObjects - 1) :
+            endElement;
         // Burn lines if not starting at first element
         for (unsigned int k = 0; k < _startRenderIdx; ++k) {
             skipSingleLineInFile(file);
@@ -269,8 +272,8 @@ void RenderableSmallBody::readDataFile(const std::string& filename) {
 void RenderableSmallBody::initializeFileReading() {
     _startRenderIdx.removeOnChange(_startRenderIdxCallbackHandle);
     _sizeRender.removeOnChange(_sizeRenderCallbackHandle);
-    _startRenderIdx.setMaxValue(_numObjects - 1);
-    _sizeRender.setMaxValue(_numObjects);
+    _startRenderIdx.setMaxValue(static_cast<unsigned int>(_numObjects - 1));
+    _sizeRender.setMaxValue(static_cast<unsigned int>(_numObjects));
     _startRenderIdx = static_cast<unsigned int>(0);
     _sizeRender = static_cast<unsigned int>(_numObjects);
     _startRenderIdxCallbackHandle = _startRenderIdx.onChange(_updateStartRenderIdxSelect);
@@ -278,7 +281,7 @@ void RenderableSmallBody::initializeFileReading() {
     // If a limit wasn't specified in dictionary, set it to # lines in file
     // minus the header line (but temporarily disable callback to avoid 2nd call)
     _upperLimit.removeOnChange(_upperLimitCallbackHandle);
-    _upperLimit.setMaxValue(_numObjects);
+    _upperLimit.setMaxValue(static_cast<unsigned int>(_numObjects));
     _upperLimit = static_cast<unsigned int>(_numObjects);
     _upperLimitCallbackHandle = _upperLimit.onChange(_reinitializeTrailBuffers);
 }
@@ -390,7 +393,9 @@ void RenderableSmallBody::readOrbitalParamsFromThisLine(bool firstDataLine,
     _data.push_back(keplerElements);
     _sbNames.push_back(name);
     const double scale = static_cast<double>(_segmentQuality) * 10.0;
-    _segmentSize.push_back(scale + (scale / pow(1 - keplerElements.eccentricity, 1.2)));
+    _segmentSize.push_back(
+        static_cast<size_t>(scale + (scale / pow(1 - keplerElements.eccentricity, 1.2)))
+    );
 }
 
 static double importAngleValue(const std::string& angle) {
